added fourFOur rhythm. Moced stuff around so we can have functions in songs

+++ build.sh
@@ -2,19 +2,26 @@
 
 
 
+songName="$1"
 
-rm -rf build
-mkdir build
+if [[ "$songName" == "" ]]; then
+  rm -rf build
+  mkdir build
+fi
 
 python ./generateMidiPorts.py build/midiPorts.m4
 
-cp "songs/$1.ck" build/_cur_song
 
-cp -r src/* build
 
+if [[ "$songName" == "" ]]; then
+  echo "" > build/_cur_song
+else
+  cp "songs/$1.ck" build/_cur_song
+fi
+
+rm -rf build/parts
 mkdir build/parts
 cp -r parts/* build/parts
-
 
 
 cd build/parts || exit
@@ -23,15 +30,23 @@ echo "" > ../_all_parts.m4
 
 while IFS= read -r -d '' file
 do
-  echo "include(parts/$file)" > ../_all_parts.m4
+  echo "include(parts/$file)" >> ../_all_parts.m4
 done < <(find . -name '*.ck' -print0)
 
 
 cd ../ || exit
 
-while IFS= read -r -d '' file
-do
-  m4 "$file" > tmp
-  mv tmp "${file%.*}.ck"
-done < <(find . -name '*.ck' -print0)
-
+if [[ "$songName" == "" ]]; then
+  cp -r ../src/* .
+  while IFS= read -r -d '' file
+  do
+    m4 "$file" > tmp
+    echo "$file"
+    mv tmp "$file"
+  done < <(find . -name '*.ck' -print0)
+else
+  cp ../src/run.ck run.ck
+  echo "Building song"
+  m4 run.ck > tmp
+  mv tmp run.ck
+fi
